requirements.txt
torch==2.0.1
transformers==4.31.0
datasets==2.14.0
pyyaml==6.0
matplotlib==3.7.2